Some crud update for item and classroom

+++ BorrowThings/src/main/java/ch/hearc/model/ClassroomRepository.java
@@ -2,6 +2,6 @@ package ch.hearc.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClassroomRepository extends JpaRepository<Classroom, Long> {
+public interface ClassroomRepository extends JpaRepository<Classroom, Integer> {
 
 }
